+++ src/components/organisms/GallerySlider/GallerySlider.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Navigation, Pagination } from 'swiper/modules';
 import { ActionButton } from '../../atoms';
-import { SCREEN_WIDTH } from '../../../constants/screenWidth';
+// import { SCREEN_WIDTH } from '../../../constants/screenWidth';
 
 import type { GallerySliderItem } from '../../../types/GallerySliderItem';
 
@@ -37,10 +37,10 @@ export const GallerySlider: React.FC<GallerySliderProps> = ({
             nextEl: `.${styles.next}`,
           }}
           pagination={{ clickable: true }}
-          breakpoints={{
-            [SCREEN_WIDTH.SLIDER_TABLET]: { slidesPerView: 1 },
-            [SCREEN_WIDTH.SLIDER_DESKTOP]: { slidesPerView: 1 },
-          }}
+          // breakpoints={{
+          //   [SCREEN_WIDTH.SLIDER_TABLET]: { slidesPerView: 1 },
+          //   [SCREEN_WIDTH.SLIDER_DESKTOP]: { slidesPerView: 1 },
+          // }}
         >
           {gallerySlides.map((slide, index) => (
             <SwiperSlide key={index}>
