--- src/services/staticGallerySlides.ts
@@ -1,19 +1,0 @@
-import type { GallerySliderItem } from '../types/GallerySliderItem';
-
-export const staticGallerySlides: GallerySliderItem[] = [
-  {
-    imgUrl: '/img/phones/apple-iphone-14-slider/slider-image-iphone-14.webp',
-    linkUrl: '/phones/iphone-14',
-    alt: 'iPhone 14',
-  },
-  {
-    imgUrl: '/img/phones/apple-iphone-14-slider/slider-image-iphone-14.webp',
-    linkUrl: '/phones/iphone-14',
-    alt: 'iPhone 14 second',
-  },
-  {
-    imgUrl: '/img/phones/apple-iphone-14-slider/slider-image-iphone-14.webp',
-    linkUrl: '/phones/iphone-14',
-    alt: 'iPhone 14 third',
-  },
-];
