+++ src/styles/_variables.scss
@@ -3,7 +3,6 @@ $default-font-size: 16px;
 $max-container-width: 1136px;
 // Breakpoints
 $breakpoint-sm: 576px;
-$breakpoint-md-slider: 640px;
 $breakpoint-md: 744px;
 $breakpoint-lg: 992px;
 $breakpoint-xl: 1200px;
